% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_segment_glyph.R
\name{geom_segment_glyph}
\alias{geom_segment_glyph}
\title{GeomSegmentGlyph}
\usage{
geom_segment_glyph(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  ...,
  x_major = NULL,
  x_minor = NULL,
  y_major = NULL,
  y_minor = NULL,
  yend_minor = NULL,
  width = ggplot2::rel(2.1),
  height = ggplot2::rel(2.1),
  a_x = 0.15,
  a_y = 0.08,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\value{
description
}
\description{
GeomSegmentGlyph
}
