% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_segment_glyph.R
\docType{data}
\name{GeomSegmentGlyph}
\alias{GeomSegmentGlyph}
\title{GeomSegmentGlyph}
\format{
An object of class \code{GeomSegmentGlyph} (inherits from \code{GeomSegment}, \code{Geom}, \code{ggproto}, \code{gg}) of length 2.
}
\usage{
GeomSegmentGlyph
}
\value{
description
}
\description{
GeomSegmentGlyph
}
\keyword{datasets}
