% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_segment_glyph.R
\name{geom_segment_glyph}
\alias{geom_segment_glyph}
\title{GeomSegmentGlyph}
\usage{
geom_segment_glyph(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  ...,
  x_major = NULL,
  x_minor = NULL,
  y_major = NULL,
  y_minor = NULL,
  yend_minor = NULL,
  width = 0.1,
  x_scale = identity,
  y_scale = identity,
  height = 0.1,
  global_rescale = TRUE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{x_major, x_minor, y_major, y_minor, xend_minor, yend_minor}{The name of the
variable (as a string) for the major and minor x and y axes. \code{x_major}
and \code{y_major} specify a longitude and latitude on a map while
\code{x_minor}, \code{y_minor}, \code{xend_minor}, and \code{yend_minor}
provide the structure for glyph.}

\item{y_scale, x_scale}{The scaling function to be applied to each set of
minor values within a grid cell. The default is \code{\link{identity}} which
produces a result without scaling.}

\item{height, width}{The height and width of each glyph.}

\item{global_rescale}{Determines whether or not the rescaling is performed
globally or separately for each individual glyph.}
}
\value{
a ggplot object
}
\description{
Create a segment plot glyph in ggplot2
}
