% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mainland_us}
\alias{mainland_us}
\title{rnaturalearth Mainland US Map Shapefile}
\format{
\subsection{\code{mainland_us}}{

A data frame with 49 rows and 122 columns:
\describe{
\item{postal}{The state abbreviation}
\item{name_sv}{The name of the state}
\item{geometry}{The polygon representing the shape of the tsate}
...
}
}
}
\usage{
mainland_us
}
\description{
A shapefile from rnatural earth of the contiguous United States.
This shapefile was also simplified with rmapshaper using ms_simplify.
}
\keyword{datasets}
