% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_segment_glyph.R
\name{rescale01x}
\alias{rescale01x}
\title{rescale01x}
\usage{
rescale01x(x, xlim = NULL)
}
\arguments{
\item{x}{The input array to be rescaled}

\item{xlim}{The limit of the rescaling. If NULL, the range of x is used.}
}
\value{
A rescaled array based on the input array and provided limit that provides values between 0 and 1
}
\description{
Rescale x-axis to 0,1
}
