% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_segment_glyph.R
\name{rescale11x}
\alias{rescale11x}
\title{rescale11x}
\usage{
rescale11x(x, xlim = NULL)
}
\arguments{
\item{x}{The input array to be rescaled}

\item{xlim}{The limit of the rescaling. If NULL, the range of x is used.}
}
\value{
A rescaled array based on the input array and provided limit that provides values between -1 and 1
}
\description{
Rescale x-axis to -1,1
}
