% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_segment_glyph.R
\name{rescale11y}
\alias{rescale11y}
\title{rescale11y}
\usage{
rescale11y(y, yend, ylim = NULL)
}
\arguments{
\item{y, yend}{The input arrays to be rescaled}

\item{ylim}{The limit of the rescaling. If NULL, the maximum value from y and yend is used.}
}
\value{
A list of two rescaled arrays based on the input array and provided limit that provides values between -1 and 1
}
\description{
Rescale y-axis to -1,1
This rescaling works differently to x rescaling because y and y end need to be rescaled together
}
