% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_glyph.R
\name{st_glyph}
\alias{st_glyph}
\title{Create a glyph map}
\usage{
st_glyph(data)
}
\arguments{
\item{data}{Default dataset to use for plot. If not already a data.frame,
will be converted to one by \code{\link[ggplot2:fortify]{fortify()}}. If not specified,
must be supplied in each layer added to the plot.}

\item{string}{A dataframe with the fields: name, month, avgmin, and avgmax.}
}
\value{
A ggplot object.
}
\description{
Create a glyph map
}
\examples{
data <- tribble(
 ~name, ~month, ~avgmin, ~avgmax,
 "A", 1, 0.5, 1.5,
 "A", 2, 0.6, 1.6,
 "A", 3, 0.7, 1.7,
)
st_glyph(data)
}
