% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stations}
\alias{stations}
\title{NOAA Weather Station Data}
\format{
\subsection{\code{stations}}{

A data frame with 17,253 rows and 12 columns:
\describe{
\item{name}{Station Location}
\item{tmin, tmax}{The minimum and maximum temperatures}
\item{prcp}{Percepitation}
...
}
}
}
\source{
\url{https://www.ncei.noaa.gov/access/search/data-search/daily-summaries}
}
\usage{
stations
}
\description{
A subset of data from NOAA weather stations in the United States.
Selected stations include: MIAMI BEACH, MT LECONTE, RALEIGH AP,
KANSAS CITY INTL AP, MCCARRAN INTL AP, CHEYENNE, NY CITY CNTRL PARK
SANTA MONICA MUNI AP and SEATTLE TACOMA AP
}
\keyword{datasets}
